map fixes

+++ app/src/main/java/com/example/albertomariopirovano/safecar/data_comparators/KMComparator.java
@@ -11,6 +11,6 @@ import java.util.Comparator;
 public class KMComparator implements Comparator<Trip> {
 
     public int compare(Trip trip1, Trip trip2){
-        return ((Integer) trip1.km).compareTo((Integer) trip2.km);
+        return ((Float) trip1.km).compareTo((Float) trip2.km);
     }
 }
